+++ src/Core/CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Added
 
 - AWS api-change: Added `us-isob-west-1` region
+- Added `DescribeLogGroups` method
 
 ### Dependency bumped